Removed obsolete types and methods
