+++ src/AgentFramework.Core/Models/Credentials/CredentialPreviewAttributeConverter.cs
@@ -23,7 +23,7 @@ namespace AgentFramework.Core.Models.Credentials
 
             var obj = new CredentialPreviewAttribute();
             obj.Name = name.Value<string>();
-            obj.MimeType = mimeType?.Value<string>() ?? CredentialMimeTypes.TextMimeType;
+            obj.MimeType = mimeType.HasValues ? mimeType.Value<string>() : CredentialMimeTypes.TextMimeType;
             obj.Value = CredentialUtils.CastAttribute(value, obj.MimeType);
             return obj;
         }
